Moves shift_back methods into LetterShiftBack module
