+++ test/letter_shift_test.rb
@@ -30,25 +30,4 @@ class LetterShiftTest < MiniTest::Test
 
     assert_equal "urlci", @enigma.shift_all("ducks")
   end
-
-  def test_a_encrypted_letters_rotated_back
-    assert_equal "ducks", @enigma.shift_back_a("uucki")
-  end
-
-  def test_b_encrypted_letters_rotated_back
-    assert_equal "ducks", @enigma.shift_back_b("drcks")
-  end
-
-  def test_c_encrypted_letters_rotated_back
-    assert_equal "ducks", @enigma.shift_back_c("dulks")
-  end
-
-  def test_d_encrypted_letters_rotated_back
-    assert_equal "ducks", @enigma.shift_back_d("duccs")
-  end
-
-  def test_all_chars_are_rotated_back
-    assert_equal "ducks", @enigma.shift_all_back("urlci")
-
-  end
 end
