+++ test/test_helper.rb
@@ -9,4 +9,5 @@ require "pry"
 require "./lib/shifts"
 require "./lib/enigma"
 require "./lib/letter_shift"
+require "./lib/letter_shift_back"
 require "./lib/encrypter" #delete later
